Ticket 49746 - Segfault during replication startup on Arm device
Bug Description:  There was an inproper use of format strings and int
                  types in agmt_update_maxcsn().  This was addressed in
                  ticket 49618 in 1.4.0, along with other fixes for
                  compiler warnings.

Fix Description:  Backport ticket 49618 to 1.3.8.

https://pagure.io/389-ds-base/issue/49746

+++ ldap/servers/slapd/abandon.c
@@ -148,7 +148,7 @@ do_abandon(Slapi_PBlock *pb)
         slapi_log_access(LDAP_DEBUG_STATS, "conn=%" PRIu64 " op=%d ABANDON"
                                            " targetop=%d msgid=%d nentries=%d etime=%" PRId64 ".%010" PRId64 "\n",
                          pb_conn->c_connid, pb_op->o_opid, o->o_opid, id,
-                         o->o_results.r.r_search.nentries, o_hr_time_end.tv_sec, o_hr_time_end.tv_nsec);
+                         o->o_results.r.r_search.nentries, (int64_t)o_hr_time_end.tv_sec, (int64_t)o_hr_time_end.tv_nsec);
     }
 
     PR_ExitMonitor(pb_conn->c_mutex);
